pyproject.toml
[project]
name = "newsletter-content-generator"
version = "0.1.0"
description = "A macOS application that aggregates tech newsletters, synthesizes content using an LLM, and generates blog posts and TikTok scripts saved to Apple Notes"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Newsletter Content Generator Team" }
]
keywords = ["newsletter", "content-generation", "llm", "apple-notes", "tiktok", "blog"]

dependencies = [
    "pyyaml>=6.0",
    "feedparser>=6.0",
    "beautifulsoup4>=4.12",
    "openai>=1.0",
    "macnotesapp>=0.7",
    "lxml>=5.0",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.100",
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-timeout>=2.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
    "types-beautifulsoup4>=4.12",
]

[project.scripts]
newsletter-generator = "newsletter_generator.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/newsletter_generator"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "property: marks tests as property-based tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.hypothesis]
# Hypothesis settings for property-based testing
# Minimum 100 examples as specified in the design document

[tool.hypothesis.profiles.default]
max_examples = 100
deadline = 5000
suppress_health_check = ["too_slow"]

[tool.hypothesis.profiles.ci]
max_examples = 200
deadline = 10000

[tool.hypothesis.profiles.dev]
max_examples = 50
deadline = 3000

[tool.coverage.run]
source = ["src/newsletter_generator"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
show_missing = true

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["newsletter_generator"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
namespace_packages = true
explicit_package_bases = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = [
    "feedparser.*",
    "macnotesapp.*",
]
ignore_missing_imports = true
